r24_pwa/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.3
idna==2.8
py==1.9.0
pytz==2019.1
requests==2.22.0
retry==0.9.2
sqlparse==0.3.1
urllib3==1.25.10
uWSGI==2.0.17
google-cloud-datastore==2.1.0